refactor: improving the code for getting information from a service with datasets
