+++ services/agents/app/services/data/__init__.py
@@ -1,37 +1,24 @@
-import requests
-from functools import wraps
 from crewai.tools import tool
 
-from .tools import get_sample_sizes_fot_all_data
-from app.config import config_url
+from .utils import get_json, handle_errors, get_sample_sizes_for_all_data
+from app.logs import get_logger
 
-def handle_errors(func):
-    """Декоратор для обработки ошибок API запросов"""
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        try:
-            return func(*args, **kwargs)
-        except requests.exceptions.Timeout:
-            return {"ERROR": f"Timeout: {func.__name__} не ответил..."}
-        except requests.exceptions.ConnectionError:
-            return {"ERROR": f"Connection error: Не удалось подключиться к {config_url.DMS_URL}"}
-        except requests.exceptions.HTTPError as e:
-            return {"ERROR": f"HTTP error {e.response.status_code}: {e.response.text}"}
-        except Exception as e:
-            return {"ERROR": f"Ошибка в {func.__name__}: {str(e)}"}
-    return wrapper
-
+logger = get_logger(__name__)
 
 @tool("GetDatasetInfo")
 @handle_errors
 def get_dataset_info(dataset_id: str) -> dict:
-    """Получить информацию о датасете по ID"""
-    try:
-        resp = requests.get(f"{config_url.DMS_URL}/api/datasets/{dataset_id}")
-        data = resp.json()
-        return data
-    except Exception as e:
-        return {"ERROR": f"Ошибка: {e}"}
+    """
+    Получить всю информацию о датасете по ID
+    
+    Args:
+        dataset_id: Id датасета
+        
+    Returns:
+        Dict с информацией о датасете или ошибкой
+    """    
+    return get_json(f"/api/datasets/{dataset_id}")
+
     
 @tool("GetDatasetVersionInfo")
 @handle_errors
@@ -39,13 +26,18 @@ def get_version_info(
     dataset_id: str,
     version_id: str
 ) -> dict:
-    """Получить информацию о версии датасета по ID"""
-    try:
-        resp = requests.get(f"{config_url.DMS_URL}/api/datasets/{dataset_id}/versions/{version_id}")
-        data = resp.json()
-        return data
-    except Exception as e:
-        return {"ERROR": f"Ошибка: {e}"}
+    """
+    Получить информацию о версии датасета по ID
+    
+    Args:
+        dataset_id: Id датасета
+        version_id: Id версии
+        
+    Returns:
+        Dict с информацией о версии или ошибкой
+    """
+    return get_json(f"/api/datasets/{dataset_id}/versions/{version_id}")
+
     
 @tool("GetDatasetVersionSplitInfo")
 @handle_errors
@@ -53,22 +45,22 @@ def get_version_split_info(
     dataset_id: str,
     version_id: str
 ) -> dict:
-    """Получить информацию о версии датасета по ID"""
-    try:
-        resp = requests.get(f"{config_url.DMS_URL}/api/datasets/{dataset_id}/versions/{version_id}")
-        data = resp.json()
+    """
+    Получить информацию о разбиении датасета (train/val/test)
+    
+    Args:
+        dataset_id: Id датасета
+        version_id: Id версии
+        
+    Returns:
+        Dict с информацией о разбиении или ошибкой
+    """
+    json = get_json(f"/api/datasets/{dataset_id}/versions/{version_id}")
+    return get_sample_sizes_for_all_data(json)
 
-        return get_sample_sizes_fot_all_data(data)
-    except Exception as e:
-        return {"ERROR": f"Ошибка: {e}"}
 
 @tool("ListDatasets")
 @handle_errors
 def list_datasets() -> dict:
     """Список всех датасетов"""
-    try:
-        resp = requests.get(f"{config_url.DMS_URL}/api/datasets/")
-        datasets = resp.json()
-        return datasets
-    except Exception as e:
-        return {"ERROR": f"Ошибка: {e}"}
+    return get_json("/api/datasets/")
